backend/requirements.txt
# backend/requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
SQLAlchemy==2.0.29
Werkzeug==2.3.8
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.1
Flask-Mail==0.9.1
cryptography==42.0.5 # <-- 新增：用於 MySQL 8.0 的 caching_sha2_password 認證方法
# gunicorn==22.0.0
